+++ .gitignore
@@ -8,6 +8,9 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
+# build artifacts
+src/**/*.css
+
 # Runtime data
 pids
 *.pid
