+++ package.json
@@ -25,7 +25,7 @@
     "start-pc": "set PORT=3000 && npm-run-all -p watch-css start-js-pc",
     "start-pc-local": "set PORT=8080 && set REACT_APP_LOCAL=1 && npm-run-all -p watch-css start-js-pc",
     "build": "npm run build-css && react-scripts build",
-    "test": "react-scripts test --forceExit --env=jsdom",
+    "test": "npm run build-css && react-scripts test --forceExit --env=jsdom",
     "eject": "react-scripts eject",
     "lint": "eslint src/",
     "lintfix": "eslint src/ --fix"
