--- src/components/css/App.css
@@ -1,10 +1,0 @@
-@import 'GameScreen.css';
-@import 'GameButton.css';
-
-.App{
-    /*
-    The max width lets the app keep its mobile appearance even on larger screen,
-    its mainly a short fix since we dont have a PC version of the screens yet.
-     */
-    /*max-width: 400px;*/
-}
