--- src/components/css/GameButton.css
@@ -1,11 +1,0 @@
-.GameButton {
-    height: 100%;
-    width: 100%;
-    min-height: 50px;
-}
-
-.GameButtonContainer {
-    height: auto;
-    width: auto;
-
-}
