--- src/components/css/GameScreen.css
@@ -1,8 +1,0 @@
-.GameScreen {
-    position: fixed;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    right: 0;
-    overflow: auto;
-}
