--- src/components/stylesheets/Component.css
@@ -1,67 +1,0 @@
-.mainScreen {
-  width: 100%;
-  font-size: 30px;
-  height: 100vh; }
-
-.fluidGridList {
-  flex-wrap: 'nowrap';
-  overflow: auto;
-  white-space: nowrap;
-  transform: 'translateZ(0)'; }
-
-.fluidGridItem {
-  display: inline-block; }
-
-.iconListItemSize {
-  width: 64px;
-  height: 64px; }
-
-.colorListItemSize {
-  width: 64px;
-  height: 64px; }
-
-.tabClass {
-  color: black; }
-
-.colorPicker {
-  margin-top: 10px; }
-
-.iconPreviewWindow {
-  border-radius: 170px;
-  width: 160px;
-  height: 160px;
-  display: block;
-  margin: 0px auto;
-  text-align: center; }
-
-.center {
-  display: block;
-  margin: 0 auto;
-  margin-top: 30px;
-  margin-left: 30px;
-  width: 100px;
-  height: 100px; }
-
-.buttonContainer {
-  flex-wrap: nowrap;
-  margin: 'auto';
-  width: '100%';
-  margin-top: 0px; }
-
-.button {
-  display: block;
-  width: 100%; }
-
-.sessionContainer {
-  flex-wrap: nowrap;
-  margin: 'auto';
-  width: '100%';
-  margin-top: 0px; }
-
-.sessionHeader {
-  flex-wrap: nowrap;
-  margin: 'auto';
-  width: '100%';
-  margin-top: 0px;
-  background-color: #2196f3;
-  color: antiquewhite; }
